+++ model/common/tests/metric_tests/test_metric_scalars.py
@@ -14,7 +14,7 @@
 import numpy as np
 import pytest
 from gt4py.next import as_field
-from gt4py.next.ffront.fbuiltins import int32, int64
+from gt4py.next.ffront.fbuiltins import int32
 
 from icon4py.model.common.dimension import KDim
 from icon4py.model.common.metrics.metric_scalars import compute_kstart_dd3d, compute_scalfac_dd3d
@@ -29,10 +29,10 @@ def test_compute_dd3d(icon_grid, metrics_savepoint, grid_savepoint, backend):
     scalfac_dd3d_ref = metrics_savepoint.scalfac_dd3d()
     kstart_dd3d_ref = 2  # TODO: ref value not serialized
     scalfac_dd3d_full = zero_field(icon_grid, KDim)
-    k_index = as_field((KDim,), np.arange(icon_grid.num_levels, dtype=int64))
+    k_index = as_field((KDim,), np.arange(icon_grid.num_levels, dtype=int32))
     divdamp_trans_start = 12500.0
     divdamp_trans_end = 17500.0
-    divdamp_type = int64(3)
+    divdamp_type = 3
 
     compute_scalfac_dd3d.with_backend(backend=backend)(
         vct_a=grid_savepoint.vct_a(),
